1.6 - cafe SPA for customer feedback
initial files for exercise 1.6

The student restaurant of the University of Helsinki 'Unicafe' collects feedback from its customers. Task is to implement a web application for collecting customer feedback. There are only three options for feedback: good, neutral, and bad.

The application must display the total number of collected feedback for each category. Application needs to work only during a single browser session. Once the page is refreshed, the collected feedback is allowed to disappear.